♻️Refactor: 공용 컴포넌트 리펙토링 진행
