+++ app/community/post/[postId]/page.tsx
@@ -5,7 +5,7 @@ import Link from "next/link";
 import { getServerSession } from "next-auth";
 import { authOption } from "@/app/api/auth/[...nextauth]/route";
 import Comment from "@/components/Comment";
-import DeleteBtn from "@/app/community/post/[postId]/DeleteBtn";
+import DeleteBtn from "@/components/Button/DeleteBtn";
 
 const Editor = dynamic(() => import("@/app/community/new/Editor"), {
   ssr: false,
@@ -29,6 +29,7 @@ const getPost = async (id: string) =>
     async (res) => await res.json()
   );
 
+//TODO views의 업데이트는 post를 불러오는 단계에서 업데이트 하도록 변경 예정
 const updateViews = async (id: string) =>
   await fetch(`${baseURL}/api/post/${id}/views`, { cache: "no-cache" });
