+++ components/Divder.tsx
@@ -1,0 +1,2 @@
+const Divider = () => <span className="divider" />;
+export default Divider;
